chore: improve project structure
