+++ client/src/app/(auth)/login/page.tsx
@@ -1,11 +1,11 @@
-import LoginAccount from '@/features/login/login';
+import LoginPageComponent from '@/features/login/login';
 import { DotBackground } from '@/shared/components/background';
 
 export default function LoginPage() {
   return (
     <main>
       <DotBackground />
-      <LoginAccount />
+      <LoginPageComponent />
     </main>
   );
 }
