+++ client/src/features/student-profile/components/info-box.tsx
