+++ client/src/features/students/components/delete-student.tsx
